requirements.txt
streamlit==1.41.1
pypdf==5.1.0
sentence-transformers==3.3.1
faiss-cpu==1.12.0
numpy==1.26.4
python-dotenv==1.0.1
openai==1.57.0